Enable strict concurrency checking, support only collection views
Table view, stack view, and fetched results controller have not been used or tested for a long time so there’s no need to retain them.

Lots of tests are failing but that might be because of changes in iOS; the collection view does not request cells in the same way it did in previous versions.