Log section addresses, only log in debug
We often see many sections of the same type, which makes it quite hard to distinguish them. Showing the addresses makes it a bit easier to track.

`debugLog(_:)` is not a particularly expensive call, but we do call it a lot. I also want the new `Section.debugDescription` property to only be used in debug builds so it’s best to just wrap all these symbols in debug checks.